[test] Simplify headers for idempotent
