+++ tests/src/when-component-idempotent-header.cpp
@@ -4,7 +4,7 @@
  * SPDX - License - Identifier: MIT
  */
 
-#include "when-component-idempotent-common.hpp"
+#include "idempotent.hpp"
 
 #include "pbdt/bdd.hpp"
 
@@ -37,7 +37,8 @@ struct When
 
 int main()
 {
-    idempotent<ToContainer, When>();
+    const AcceptableRawContext<CompletableRawWhenContext<When>> acceptable;
+    acceptable.accept(IdempotentValidator<When, ToContainer>{});
 
     return EXIT_SUCCESS;
 }
